Struct black76::Inputs

pub struct Inputs {
    pub option_type: OptionType,
    pub f: f32,
    pub k: f32,
    pub p: Option<f32>,
    pub r: f32,
    pub q: f32,
    pub t: f32,
    pub sigma: Option<f32>,
}

The inputs to the Black-Scholes-Merton model.

Fields
option_type: OptionType

The type of the option (call or put)

f: f32

Futures price

k: f32

Strike price

p: Option<f32>

Option price

r: f32

Risk-free rate

q: f32

Dividend yield

t: f32

Time to maturity in years

sigma: Option<f32>

Volatility

Implementations

impl Inputs

Methods for calculating the price, greeks, and implied volatility of an option.

pub fn new(
option_type: OptionType,
f: f32,
k: f32,
p: Option<f32>,
r: f32,
q: f32,
t: f32,
sigma: Option<f32>
) -> Self

Creates instance ot the Inputs struct.

Arguments

option_type - The type of option to be priced.

f - The current price of the underlying asset.

k - The strike price of the option.

p - The dividend yield of the underlying asset.

r - The risk-free interest rate.

q - The dividend yield of the underlying asset.

t - The time to maturity of the option in years.

sigma - The volatility of the underlying asset.

Example

use blackscholes::{Inputs, OptionType};
let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));

Returns

An instance of the Inputs struct.